separate features for sign up in a new module
