+++ saas_portal/controllers/main.py
@@ -4,12 +4,10 @@ from openerp import SUPERUSER_ID
 from openerp.addons.web import http
 from openerp.addons.web.http import request
 from openerp.addons.auth_oauth.controllers import main as oauth
-from openerp.addons import auth_signup
 import werkzeug
 import simplejson
 import uuid
 import random
-import re
 
 
 class SignupError(Exception):
@@ -32,18 +30,17 @@ class SaasPortal(http.Controller):
         dbtemplate = self.get_template()
         client_id = self.get_new_client_id(full_dbname)
         request.registry['oauth.application'].create(request.cr, SUPERUSER_ID, {'client_id': client_id, 'name':full_dbname})
-        user = self.update_user_and_partner(full_dbname)
+        organization = self.update_user_and_partner(full_dbname)
         params = {
             'scope': 'userinfo force_login trial skiptheuse',
             'state': simplejson.dumps({
                 'd': full_dbname,
                 'u': '%s://%s' % (scheme, full_dbname.replace('_', '.')),
-                'o': user.organization,
+                'o': organization,
                 'db_template': dbtemplate,
             }),
             'redirect_uri': '{scheme}://{saas_server}/saas_server/new_database'.format(scheme=scheme, saas_server=saas_server),
             'response_type': 'token',
-            'organization': post.get('organization'),
             'client_id': client_id,
         }
         return request.redirect('/oauth2/auth?%s' % werkzeug.url_encode(params))
@@ -113,6 +110,7 @@ class SaasPortal(http.Controller):
         user_model = request.registry.get('res.users')
         user = user_model.browse(request.cr, SUPERUSER_ID, request.uid)
         partner_model = request.registry.get('res.partner')
+        organization = user.organization or database.split('_')[0]
         wals = {
             'name': user.organization,
             'is_company': True,
@@ -131,7 +129,7 @@ class SaasPortal(http.Controller):
             'parent_id': pid
         }
         user_model.write(request.cr, SUPERUSER_ID, user.id, vals)
-        return user
+        return organization
 
 
 class OAuthLogin(oauth.OAuthLogin):
@@ -144,15 +142,8 @@ class OAuthLogin(oauth.OAuthLogin):
             fields = ['share', 'database']
             data = user.search_read(request.cr, SUPERUSER_ID, domain, fields)
             if data and data[0]['share'] and data[0]['database']:
-                kw['redirect'] = '/saas_server/tenant'
+                kw['redirect'] = '/saas_portal/tenant'
         return super(OAuthLogin, self).web_login(*args, **kw)
-
-    @http.route()
-    def web_auth_signup(self, *args, **kw):
-        if kw.get('dbname', False):
-            redirect = '/saas_portal/book_then_signup'
-            kw['redirect'] = '%s?dbname=%s' % (redirect, kw['dbname'])
-        return super(OAuthLogin, self).web_auth_signup(*args, **kw)
 
     @http.route()
     def web_auth_reset_password(self, *args, **kw):
@@ -163,51 +154,5 @@ class OAuthLogin(oauth.OAuthLogin):
             fields = ['share', 'database']
             data = user.search_read(request.cr, SUPERUSER_ID, domain, fields)
             if data and data[0]['share'] and data[0]['database']:
-                kw['redirect'] = '/saas_server/tenant'
+                kw['redirect'] = '/saas_portal/tenant'
         return super(OAuthLogin, self).web_auth_reset_password(*args, **kw)
-
-
-class AuthSignupHome(auth_signup.controllers.main.AuthSignupHome):
-
-    def get_auth_signup_qcontext(self):
-        qcontext = super(AuthSignupHome, self).get_auth_signup_qcontext()
-        context = request.context
-        if not qcontext.get('plans', False):
-            sp = request.registry.get('saas_server.plan')
-            plan_ids = sp.search(request.cr, SUPERUSER_ID, [], context=context)
-            qcontext['plans'] = sp.browse(request.cr, SUPERUSER_ID, plan_ids, context=context)
-        if not qcontext.get('countries', False):
-            orm_country = request.registry.get('res.country')
-            country_ids = orm_country.search(request.cr, SUPERUSER_ID, [], context=context)
-            countries = orm_country.browse(request.cr, SUPERUSER_ID, country_ids, context=context)
-            qcontext['countries'] = countries
-        if not qcontext.get('base_saas_domain', False):
-            qcontext['base_saas_domain'] = self.get_saas_domain()
-        return qcontext
-
-    def get_saas_domain(self):
-        config = request.registry['ir.config_parameter']
-        full_param = 'saas_portal.base_saas_domain'
-        base_saas_domain = config.get_param(request.cr, SUPERUSER_ID, full_param)
-        return base_saas_domain
-
-    def do_signup(self, qcontext):
-        values = dict((key, qcontext.get(key)) for key in ('login', 'name', 'password'))
-        values['email'] = qcontext['login']
-        if qcontext.get('plan_id', False):
-            values['plan_id'] = qcontext['plan_id']
-        if qcontext.get('organization', False):
-            values['organization'] = qcontext['organization']
-        if qcontext.get('country_id', False):
-            values['country_id'] = qcontext['country_id']
-        if qcontext.get('dbname', False):
-            f_dbname = '%s.%s' % (qcontext['dbname'], self.get_saas_domain())
-            full_dbname = f_dbname.replace('www.', '').replace('.', '_')
-            db_exists = openerp.service.db.exp_db_exist(full_dbname)
-            assert re.match('[a-zA-Z0-9_.-]+$', qcontext.get('dbname')), "Only letters or numbers are allowed in domain."
-            assert db_exists == False, "Domain exists"
-        assert any([k for k in values.values()]), "The form was not properly filled in."
-        assert values.get('password') == qcontext.get('confirm_password'), "Passwords do not match; please retype them."
-        assert qcontext.get('organization') != qcontext.get('name'), "Name and Organization must be different."
-        self._signup_with_values(qcontext.get('token'), values)
-        request.cr.commit()
