+++ saas_portal_signup/controllers/__init__.py
@@ -1,0 +1,1 @@
+import main
