+++ saas_portal_signup/views/signup.xml
@@ -46,11 +46,5 @@
 			</xpath>
 		</template>
 		
-		<template id="saas_server.tenant" inherit_id="website.layout" >
-			<xpath expr="//ul[@role='menu']/li/a[@href='/web']" position="replace">
-				<a href="/saas_portal/tenant" role="menuitem">Workspace</a>
-			</xpath>
-		</template>
-		
 	</data>
 </openerp>
